requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
yfinance
anthropic
stripe
python-dotenv
gunicorn
